feat: add low stock products widget

+++ src/InventoryPlatform/InventoryPlatform.Application/DTOs/Dashboard/LowStockProductDto.cs
@@ -12,5 +12,5 @@ public sealed class LowStockProductDto
 
     public string CategoryName { get; init; } = string.Empty;
 
-    public int QuantityOnHand { get; init; }
+    public decimal QuantityOnHand { get; init; }
 }
